requirements.txt
Flask
Gunicorn
bcrypt
python-dotenv
Pillow
qrcode
requests
